Refactored this a bit to create a new component SQLInputField that contains both the CodeTextArea along with the Test Query button and all of it's logic.
