+++ packages/front-end/components/Settings/EditDataSource/ExperimentAssignmentQueries/AddEditExperimentAssignmentQueryModal.tsx
@@ -1,4 +1,4 @@
-import React, { FC, ReactElement, useEffect, useMemo, useState } from "react";
+import React, { FC, useMemo } from "react";
 import Modal from "../../../Modal";
 import {
   DataSourceInterfaceWithParams,
@@ -8,14 +8,8 @@ import { useForm } from "react-hook-form";
 import cloneDeep from "lodash/cloneDeep";
 import uniqId from "uniqid";
 import Field from "../../../Forms/Field";
-import CodeTextArea from "../../../Forms/CodeTextArea";
-import Tooltip from "../../../Tooltip/Tooltip";
 import StringArrayField from "../../../Forms/StringArrayField";
-import { validateSQL } from "../../../../services/datasources";
-import { useAuth } from "../../../../services/auth";
-import { FaPlay } from "react-icons/fa";
-import DisplayTestQueryResults from "../../DisplayTestQueryResults";
-import { TestQueryResults } from "back-end/src/types/Integration";
+import SQLInputField from "../../../../services/SQLInputField";
 
 type EditExperimentAssignmentQueryProps = {
   exposureQuery?: ExposureQuery;
@@ -32,12 +26,6 @@ export const AddEditExperimentAssignmentQueryModal: FC<EditExperimentAssignmentQ
   onSave,
   onCancel,
 }) => {
-  const [
-    testQueryResults,
-    setTestQueryResults,
-  ] = useState<TestQueryResults | null>(null);
-  const [suggestions, setSuggestions] = useState<ReactElement[]>([]);
-  const { apiCall } = useAuth();
   const modalTitle =
     mode === "add"
       ? "Add an Experiment Assignment query"
@@ -96,115 +84,11 @@ export const AddEditExperimentAssignmentQueryModal: FC<EditExperimentAssignmentQ
     ]);
   }, [userEnteredUserIdType, userEnteredDimensions, userEnteredHasNameCol]);
 
-  const handleTestQuery = async () => {
-    setTestQueryResults(null);
-    try {
-      validateSQL(userEnteredQuery, [...requiredColumns]);
-
-      const res: TestQueryResults = await apiCall("/query/test", {
-        method: "POST",
-        body: JSON.stringify({
-          query: userEnteredQuery,
-          datasourceId: dataSource.id,
-        }),
-      });
-
-      setTestQueryResults(res);
-    } catch (e) {
-      setTestQueryResults({ error: e.message });
-    }
-  };
-
   const identityTypes = useMemo(() => dataSource.settings.userIdTypes || [], [
     dataSource.settings.userIdTypes,
   ]);
 
   const saveEnabled = !!userEnteredUserIdType && !!userEnteredQuery;
-
-  useEffect(() => {
-    const result = testQueryResults?.results?.[0];
-    if (!result) return;
-
-    const suggestions: ReactElement[] = [];
-
-    const namedCols = ["experiment_name", "variation_name"];
-    const userIdTypes = identityTypes.map((type) => type.userIdType || []);
-
-    const returnedColumns = new Set<string>(Object.keys(result));
-    const optionalColumns = [...returnedColumns].filter(
-      (col) =>
-        !requiredColumns.has(col) &&
-        !namedCols.includes(col) &&
-        !userIdTypes.includes(col)
-    );
-
-    // Check if `hasNameCol` should be enabled
-    if (!userEnteredHasNameCol) {
-      // Selected both required columns, turn on `hasNameCol` automatically
-      if (
-        returnedColumns.has("experiment_name") &&
-        returnedColumns.has("variation_name")
-      ) {
-        form.setValue("hasNameCol", true);
-      }
-      // Only selected `experiment_name`, add warning
-      else if (returnedColumns.has("experiment_name")) {
-        suggestions.push(
-          <>
-            Add <code>variation_name</code> to your SELECT clause to enable
-            GrowthBook to populate names automatically.
-          </>
-        );
-      }
-      // Only selected `variation_name`, add warning
-      else if (returnedColumns.has("variation_name")) {
-        suggestions.push(
-          <>
-            Add <code>experiment_name</code> to your SELECT clause to enable
-            GrowthBook to populate names automatically.
-          </>
-        );
-      }
-    }
-
-    // Prompt to add optional columns as dimensions
-    if (optionalColumns.length > 0) {
-      suggestions.push(
-        <>
-          The following columns were returned, but will be ignored. Add them as
-          dimensions or disregard this message.
-          <ul className="mb-0 pb-0">
-            {optionalColumns.map((col) => (
-              <li key={col}>
-                <code>{col}</code> -{" "}
-                <a
-                  href="#"
-                  onClick={(e) => {
-                    e.preventDefault();
-                    form.setValue("dimensions", [
-                      ...userEnteredDimensions,
-                      col,
-                    ]);
-                  }}
-                >
-                  add as dimension
-                </a>
-              </li>
-            ))}
-          </ul>
-        </>
-      );
-    }
-
-    setSuggestions(suggestions);
-  }, [
-    requiredColumns,
-    testQueryResults,
-    userEnteredDimensions,
-    identityTypes,
-    userEnteredHasNameCol,
-    form,
-  ]);
 
   if (!exposureQuery && mode === "edit") {
     console.error(
@@ -224,9 +108,9 @@ export const AddEditExperimentAssignmentQueryModal: FC<EditExperimentAssignmentQ
       ctaEnabled={saveEnabled}
       autoFocusSelector="#id-modal-identify-joins-heading"
     >
-      <div className="my-2 ml-3">
+      <div className="my-2 ml-3 mr-3">
         <div className="row">
-          <div className="col-xs-12">
+          <div className="col-md-12 col-lg-12">
             <Field label="Display Name" required {...form.register("name")} />
             <Field
               label="Description (optional)"
@@ -247,99 +131,53 @@ export const AddEditExperimentAssignmentQueryModal: FC<EditExperimentAssignmentQ
                 form.setValue("dimensions", dimensions);
               }}
             />
-            <div className="row">
-              <div className="col-lg-8 col-md-7">
-                <label className="font-weight-bold mb-1">SQL Query</label>
-                <div>
-                  <div className="d-flex justify-content-between align-items-center p-1 border rounded">
-                    <button
-                      className="btn btn-sm btn-primary m-1"
-                      onClick={(e) => {
-                        e.preventDefault();
-                        handleTestQuery();
-                      }}
-                    >
-                      <span className="pr-2">
-                        <FaPlay />
-                      </span>
-                      Test Query
-                    </button>
-                    <div className="d-flex m-1">
-                      <label
-                        className="mr-2 mb-0"
-                        htmlFor="exposure-query-toggle"
-                      >
-                        Use Name Columns
-                      </label>
-                      <input
-                        type="checkbox"
-                        id="exposure-query-toggle"
-                        className="form-check-input "
-                        {...form.register("hasNameCol")}
-                      />
-                      <Tooltip body="Enable this if you store experiment/variation names as well as ids in your table" />
-                    </div>
-                  </div>
-                  <CodeTextArea
-                    required
-                    language="sql"
-                    value={userEnteredQuery}
-                    setValue={(sql) => form.setValue("query", sql)}
-                  />
-                  {testQueryResults && (
-                    <DisplayTestQueryResults
-                      duration={parseInt(testQueryResults.duration || "0")}
-                      requiredColumns={[...requiredColumns]}
-                      result={testQueryResults.results?.[0]}
-                      suggestions={suggestions}
-                      error={testQueryResults.error}
-                      sql={testQueryResults.sql}
-                    />
-                  )}
-                </div>
-              </div>
-              <div className="col-md-5 col-lg-4">
-                <div className="pt-md-4">
-                  <strong>Required columns</strong>
-                </div>
-                <ul>
+            <label>Required Columns</label>
+            <ul>
+              <li>
+                <code>{form.watch("userIdType")}</code>
+              </li>
+              <li>
+                <code>timestamp</code>
+              </li>
+              <li>
+                <code>experiment_id</code>
+              </li>
+              <li>
+                <code>variation_id</code>
+              </li>
+              {userEnteredHasNameCol && (
+                <>
                   <li>
-                    <code>{userEnteredUserIdType}</code>
+                    <code>experiment_name</code>
                   </li>
                   <li>
-                    <code>timestamp</code>
+                    <code>variation_name</code>
                   </li>
-                  <li>
-                    <code>experiment_id</code>
-                  </li>
-                  <li>
-                    <code>variation_id</code>
-                  </li>
-                  {userEnteredHasNameCol && (
-                    <>
-                      <li>
-                        <code>experiment_name</code>
-                      </li>
-                      <li>
-                        <code>variation_name</code>
-                      </li>
-                    </>
-                  )}
-                  {userEnteredDimensions &&
-                    userEnteredDimensions.map((dimension) => {
-                      return (
-                        <li key={dimension}>
-                          <code>{dimension}</code>
-                        </li>
-                      );
-                    })}
-                </ul>
+                </>
+              )}
+              {userEnteredDimensions &&
+                userEnteredDimensions.map((dimension) => {
+                  return (
+                    <li key={dimension}>
+                      <code>{dimension}</code>
+                    </li>
+                  );
+                })}
+            </ul>
+            <SQLInputField
+              userEnteredQuery={userEnteredQuery}
+              datasourceId={dataSource.id}
+              form={form}
+              requiredColumns={requiredColumns}
+              identityTypes={identityTypes}
+              queryType="experiment-assignment"
+              helpText={
                 <div>
                   Any additional columns you select can be listed as dimensions
                   to drill down into experiment results.
                 </div>
-              </div>
-            </div>
+              }
+            />
           </div>
         </div>
       </div>
